refactor: extract SupportCreator component from EntryContent
