+++ apps/renderer/src/modules/entry-content/index.tsx
@@ -16,14 +16,11 @@ import { useUISettingKey } from "~/atoms/settings/ui"
 import { m } from "~/components/common/Motion"
 import { ShadowDOM } from "~/components/common/ShadowDOM"
 import { AutoResizeHeight } from "~/components/ui/auto-resize-height"
-import { Button } from "~/components/ui/button"
-import { Divider } from "~/components/ui/divider"
 import { HTML } from "~/components/ui/markdown"
 import { Toc } from "~/components/ui/markdown/components/Toc"
 import { useInPeekModal } from "~/components/ui/modal/inspire/PeekModal"
 import { RootPortal } from "~/components/ui/portal"
 import { ScrollArea } from "~/components/ui/scroll-area"
-import { UserAvatar } from "~/components/user-button"
 import { isWebBuild, ROUTE_FEED_PENDING } from "~/constants"
 import { shortcuts } from "~/constants/shortcuts"
 import { useEntryReadabilityToggle } from "~/hooks/biz/useEntryActions"
@@ -45,10 +42,10 @@ import { useFeedById, useFeedHeaderTitle } from "~/store/feed"
 
 import { LoadingWithIcon } from "../../components/ui/loading"
 import { EntryPlaceholderDaily } from "../ai/ai-daily/EntryPlaceholderDaily"
-import { useTipModal } from "../wallet/hooks"
 import { setEntryContentScrollToTop, setEntryTitleMeta } from "./atoms"
 import { EntryPlaceholderLogo } from "./components/EntryPlaceholderLogo"
 import { EntryTitle } from "./components/EntryTitle"
+import { SupportCreator } from "./components/SupportCreator"
 import { EntryHeader } from "./header"
 import { EntryContentLoading } from "./loading"
 import { EntryContentProvider } from "./provider"
@@ -135,12 +132,6 @@ export const EntryContentRender: Component<{ entryId: string }> = ({ entryId, cl
         : undefined,
     [readerFontFamily],
   )
-
-  const openTipModal = useTipModal({
-    userId: feed?.ownerUserId ?? undefined,
-    feedId: entry?.feedId,
-    entryId,
-  })
 
   if (!entry) return null
 
@@ -235,44 +226,7 @@ export const EntryContentRender: Component<{ entryId: string }> = ({ entryId, cl
               </div>
             )}
 
-            {feed?.ownerUserId && (
-              <>
-                <Divider />
-
-                <div className="my-16 flex flex-col items-center gap-8">
-                  <UserAvatar
-                    className="w-40 flex-col gap-3 p-0"
-                    avatarClassName="size-12"
-                    userId={feed.ownerUserId}
-                    enableModal
-                  />
-                  <Button className="text-base" onClick={() => openTipModal()}>
-                    <i className="i-mgc-power-outline mr-1.5 text-lg" />
-                    {t("entry_content.support_creator")}
-                  </Button>
-                  {!!feed.tipUsers?.length && (
-                    <>
-                      <div className="text-sm text-zinc-500">
-                        {t("entry_content.support_amount", { amount: feed.tipUsers.length })}
-                      </div>
-                      <div className="flex w-fit max-w-80 flex-wrap gap-4">
-                        {feed.tipUsers?.map((user) => (
-                          <div key={user.id} className="size-8">
-                            <UserAvatar
-                              className="h-auto p-0"
-                              avatarClassName="size-8"
-                              userId={user?.id}
-                              enableModal={true}
-                              hideName={true}
-                            />
-                          </div>
-                        ))}
-                      </div>
-                    </>
-                  )}
-                </div>
-              </>
-            )}
+            {feed?.ownerUserId && <SupportCreator entryId={entryId} />}
           </article>
         </div>
       </ScrollArea.ScrollArea>
